+++ xtask/src/main.rs
@@ -7,16 +7,18 @@ mod util;
 use anyhow::Result;
 use cargo::{Cargo, CargoAction, Feature, Package};
 use clap::Parser;
-use opt::{Action, BuildOpt, ClippyOpt, DocOpt, Opt, QemuOpt};
+use opt::{Action, BuildOpt, ClippyOpt, DocOpt, MiriOpt, Opt, QemuOpt};
 use std::process::Command;
 use tempfile::TempDir;
 use util::{command_to_string, run_cmd};
+
+const NIGHTLY: &str = "nightly";
 
 fn build(opt: &BuildOpt) -> Result<()> {
     let cargo = Cargo {
         action: CargoAction::Build,
         features: Feature::more_code(),
-        nightly: true,
+        toolchain: opt.toolchain.or(NIGHTLY),
         packages: Package::all_except_xtask(),
         release: opt.build_mode.release,
         target: Some(*opt.target),
@@ -30,7 +32,7 @@ fn clippy(opt: &ClippyOpt) -> Result<()> {
     let cargo = Cargo {
         action: CargoAction::Clippy,
         features: Feature::more_code(),
-        nightly: true,
+        toolchain: opt.toolchain.or(NIGHTLY),
         packages: Package::all_except_xtask(),
         release: false,
         target: Some(*opt.target),
@@ -42,7 +44,7 @@ fn clippy(opt: &ClippyOpt) -> Result<()> {
     let cargo = Cargo {
         action: CargoAction::Clippy,
         features: Vec::new(),
-        nightly: false,
+        toolchain: None,
         packages: vec![Package::Xtask],
         release: false,
         target: None,
@@ -56,11 +58,25 @@ fn doc(opt: &DocOpt) -> Result<()> {
     let cargo = Cargo {
         action: CargoAction::Doc { open: opt.open },
         features: Feature::more_code(),
-        nightly: true,
+        toolchain: opt.toolchain.or(NIGHTLY),
         packages: Package::published(),
         release: false,
         target: None,
         warnings_as_errors: opt.warning.warnings_as_errors,
+    };
+    run_cmd(cargo.command()?)
+}
+
+/// Run unit tests and doctests under Miri.
+fn run_miri(opt: &MiriOpt) -> Result<()> {
+    let cargo = Cargo {
+        action: CargoAction::Miri,
+        features: [Feature::Exts].into(),
+        toolchain: opt.toolchain.or(NIGHTLY),
+        packages: [Package::Uefi].into(),
+        release: false,
+        target: None,
+        warnings_as_errors: false,
     };
     run_cmd(cargo.command()?)
 }
@@ -76,7 +92,7 @@ fn run_vm_tests(opt: &QemuOpt) -> Result<()> {
     let cargo = Cargo {
         action: CargoAction::Build,
         features,
-        nightly: true,
+        toolchain: opt.toolchain.or(NIGHTLY),
         packages: vec![Package::UefiTestRunner],
         release: opt.build_mode.release,
         target: Some(*opt.target),
@@ -94,7 +110,7 @@ fn run_host_tests() -> Result<()> {
     let cargo = Cargo {
         action: CargoAction::Test,
         features: vec![Feature::Exts],
-        nightly: true,
+        toolchain: Some(NIGHTLY.into()),
         // Don't test uefi-services (or the packages that depend on it)
         // as it has lang items that conflict with `std`.
         packages: vec![Package::Uefi, Package::UefiMacros, Package::Xtask],
@@ -152,6 +168,7 @@ fn main() -> Result<()> {
         Action::Build(build_opt) => build(build_opt),
         Action::Clippy(clippy_opt) => clippy(clippy_opt),
         Action::Doc(doc_opt) => doc(doc_opt),
+        Action::Miri(miri_opt) => run_miri(miri_opt),
         Action::Run(qemu_opt) => run_vm_tests(qemu_opt),
         Action::Test(_) => run_host_tests(),
         Action::TestLatestRelease(_) => test_latest_release(),