+++ xtask/src/opt.rs
@@ -22,6 +22,22 @@ impl Deref for TargetOpt {
 }
 
 #[derive(Debug, Parser)]
+pub struct ToolchainOpt {
+    /// Rust toolchain to use, e.g. "nightly-2022-02-24".
+    #[clap(long)]
+    toolchain: Option<String>,
+}
+
+impl ToolchainOpt {
+    /// Get the toolchain arg if set, otherwise use `default_toolchain`.
+    pub fn or(&self, default_toolchain: &str) -> Option<String> {
+        self.toolchain
+            .clone()
+            .or_else(|| Some(default_toolchain.to_string()))
+    }
+}
+
+#[derive(Debug, Parser)]
 pub struct BuildModeOpt {
     /// Build in release mode.
     #[clap(long)]
@@ -47,6 +63,7 @@ pub enum Action {
     Build(BuildOpt),
     Clippy(ClippyOpt),
     Doc(DocOpt),
+    Miri(MiriOpt),
     Run(QemuOpt),
     Test(TestOpt),
     TestLatestRelease(TestLatestReleaseOpt),
@@ -59,6 +76,9 @@ pub struct BuildOpt {
     pub target: TargetOpt,
 
     #[clap(flatten)]
+    pub toolchain: ToolchainOpt,
+
+    #[clap(flatten)]
     pub build_mode: BuildModeOpt,
 }
 
@@ -69,12 +89,18 @@ pub struct ClippyOpt {
     pub target: TargetOpt,
 
     #[clap(flatten)]
+    pub toolchain: ToolchainOpt,
+
+    #[clap(flatten)]
     pub warning: WarningOpt,
 }
 
 /// Build the docs for the uefi packages.
 #[derive(Debug, Parser)]
 pub struct DocOpt {
+    #[clap(flatten)]
+    pub toolchain: ToolchainOpt,
+
     /// Open the docs in a browser.
     #[clap(long)]
     pub open: bool,
@@ -83,11 +109,21 @@ pub struct DocOpt {
     pub warning: WarningOpt,
 }
 
+/// Run unit tests and doctests under Miri.
+#[derive(Debug, Parser)]
+pub struct MiriOpt {
+    #[clap(flatten)]
+    pub toolchain: ToolchainOpt,
+}
+
 /// Build uefi-test-runner and run it in QEMU.
 #[derive(Debug, Parser)]
 pub struct QemuOpt {
     #[clap(flatten)]
     pub target: TargetOpt,
+
+    #[clap(flatten)]
+    pub toolchain: ToolchainOpt,
 
     #[clap(flatten)]
     pub build_mode: BuildModeOpt,